tests/wpt/tests/webmessaging/with-options/undefined-transferable.html

<!doctype html>
<title>undefined as transferable</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  postMessage('', {transfer: undefined});
  onmessage = this.step_func(function(e) {
    assert_array_equals(e.ports, []);
    this.done();
  });
});
</script>
